+++ src/Dracodeum/Kit/Interfaces/Readonlyable.php
@@ -14,20 +14,16 @@ interface Readonlyable
 	/**
 	 * Check if is read-only.
 	 * 
-	 * @param bool $recursive [default = false]
-	 * <p>Check if it has been recursively set as read-only.</p>
 	 * @return bool
 	 * <p>Boolean <code>true</code> if is read-only.</p>
 	 */
-	public function isReadonly(bool $recursive = false): bool;
+	public function isReadonly(): bool;
 	
 	/**
 	 * Set as read-only.
 	 * 
-	 * @param bool $recursive [default = false]
-	 * <p>Set all the possible referenced subobjects as read-only recursively (if applicable).</p>
 	 * @return $this
 	 * <p>This instance, for chaining purposes.</p>
 	 */
-	public function setAsReadonly(bool $recursive = false): object;
+	public function setAsReadonly(): object;
 }