+++ src/Dracodeum/Kit/Options.php
@@ -88,14 +88,8 @@ IIntegerInstantiable, IFloatInstantiable, IStringInstantiable, ICallableInstanti
 		);
 		
 		//read-only
-		$this->addReadonlyCallback(function (bool $recursive): void {
-			//properties
+		$this->addReadonlyCallback(function (): void {
 			$this->setPropertiesAsReadonly();
-			
-			//recursive
-			if ($recursive) {
-				UType::setValueAsReadonly($this->getAll(true), $recursive);
-			}
 		});
 	}